chore(project): configure base config for tsdown and reuse it in current packages
