+++ configs/tsdown/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@aio-commerce-sdk/config-typescript/tsconfig.base.json",
+  "compilerOptions": {
+    "rootDir": "."
+  },
+
+  "include": ["tsdown.config.base.ts"]
+}
