+++ packages/aio-commerce-lib-auth/package.json
@@ -63,8 +63,7 @@
 
   "devDependencies": {
     "@aio-commerce-sdk/config-typescript": "workspace:*",
-    "@types/node": "^24.0.4",
-    "tsdown": "^0.12.9",
+    "@aio-commerce-sdk/config-tsdown": "workspace:*",
     "vitest": "^3.2.4"
   },
